Reworked immersion mode. I got the implementation down to about 5 lines, so I put it back in ForumsIndexActivity :v I'm not sure why it was so over-complicated. Complications with previous version of android, perhaps? I tested it extensively and could find no issues with this implementation. I based it on the sample at https://github.com/googlesamples/android-ImmersiveMode so it should be kosher.
